change les olds cast en new cast, et supprime les shadows variables

+++ src/main.cpp
@@ -12,7 +12,7 @@
 using namespace std;
 
 int main(){
-    srand(time(NULL));
+    srand(static_cast<unsigned int> (time(NULL)));
 
 
     //TEST
